// src/redux/createEntityClaims/createEntityClaims.types.ts
export interface Template {
  id: string
  entityClaimId: string
  templateId: string
  title: string
  description: string
  isPrivate: boolean
  minTargetClaims: number
  maxTargetClaims: number
  goal: string
  submissionStartDate: string
  submissionEndDate: string
}

export interface AgentRole {
  id: string
  entityClaimId: string
  role: string
  credential: string
  autoApprove: boolean
}

export interface Evaluation {
  entityClaimId: string
  id: string
  context: string
  contextLink: string
  evaluationAttributes: string[]
  evaluationMethodology: string
}

export interface ApprovalAttribute {
  attribute: string
  condition: string
}

export interface ApprovalCriterion {
  id: string
  entityClaimId: string
  context: string
  contextLink: string
  approvalAttributes: ApprovalAttribute[]
}

export interface EnrichmentResource {
  productId: string
  resource: string
}

export interface Enrichment {
  id: string
  entityClaimId: string
  context: string
  contextLink: string
  resources: EnrichmentResource[]
}

export interface EntityClaim {
  id: string
  template: Template
  agentRoles: {
    [id: string]: AgentRole
  }
  evaluations: {
    [id: string]: Evaluation
  }
  approvalCriteria: {
    [id: string]: ApprovalCriterion
  }
  enrichments: {
    [id: string]: Enrichment
  }
}

export interface EntityClaimItem {
  id: string
  template: Template
  agentRoles: AgentRole[]
  evaluations: Evaluation[]
  approvalCriteria: ApprovalCriterion[]
  enrichments: Enrichment[]
}